Sync ACPICA with Intel's version 20200110 (coming from 20190703).
* Allow more OwnerId's to prevent OWNER_ID_LIMIT exceptions.

* Return a Buffer object for all fields created via CreateField.

* Add AcpiDispatchGpe() function.

* Exit the interpreter before initializing objects within a newly
  loaded table.

* Add new 'Fields' debugger command.

* Make AcpiLoadTable() return a table index and add AcpiUnloadTable().

* Fix acpinames(8) which was broken for some time.

* Added "Windows 2019" _OSI string.

* Various improvements to iasl(8).

* Fix some compilation warnings, fix missing ACPI_PRINTF_LIKE.

* Update copyrights to 2020.

For detailed list, please see sys/contrib/dev/acpica/changes.txt.

+++ sys/contrib/dev/acpica/source/compiler/aslcodegen.c
@@ -8,7 +8,7 @@
  *
  * 1. Copyright Notice
  *
- * Some or all of this work - Copyright (c) 1999 - 2019, Intel Corp.
+ * Some or all of this work - Copyright (c) 1999 - 2020, Intel Corp.
  * All rights reserved.
  *
  * 2. License
@@ -618,8 +618,8 @@ CgWriteTableHeader (
         Op->Asl.AmlSubtreeLength +=
             strlen (AslGbl_ParseTreeRoot->Asl.Filename) + 3;
 
-        CvDbgPrint ("     Length: %lu\n",
-            strlen (AslGbl_ParseTreeRoot->Asl.Filename) + 3);
+        CvDbgPrint ("     Length: %u\n",
+            (UINT32) strlen (AslGbl_ParseTreeRoot->Asl.Filename) + 3);
 
         if (Op->Asl.CommentList)
         {
