+++ sys/contrib/dev/acpica/source/tools/acpisrc/asfile.c
@@ -8,7 +8,7 @@
  *
  * 1. Copyright Notice
  *
- * Some or all of this work - Copyright (c) 1999 - 2019, Intel Corp.
+ * Some or all of this work - Copyright (c) 1999 - 2020, Intel Corp.
  * All rights reserved.
  *
  * 2. License
@@ -546,7 +546,7 @@ AsConvertFile (
     {
         /* Decode the function bitmap */
 
-        switch ((1 << i) & Functions)
+        switch (((UINT32) 1 << i) & Functions)
         {
         case 0:
